Move OnTemperatureChangeEvent to Content.Shared Also fixes dependencies to point to the new location
