+++ Content.Server/Temperature/Systems/TemperatureSystem.cs
@@ -403,17 +403,3 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
         return (newHeatThreshold, newColdThreshold);
     }
 }
-
-public sealed class OnTemperatureChangeEvent : EntityEventArgs
-{
-    public float CurrentTemperature { get; }
-    public float LastTemperature { get; }
-    public float TemperatureDelta { get; }
-
-    public OnTemperatureChangeEvent(float current, float last, float delta)
-    {
-        CurrentTemperature = current;
-        LastTemperature = last;
-        TemperatureDelta = delta;
-    }
-}
