+++ Content.Shared/Temperature/Systems/SharedTemperatureSystem.cs
@@ -6,6 +6,11 @@ namespace Content.Shared.Temperature.Systems;
 
 public abstract partial class SharedTemperatureSystem : EntitySystem
 {
+    public override void Initialize()
+    {
+        base.Initialize();
+    }
+
     public float GetHeatCapacity(EntityUid uid, TemperatureComponent? comp = null, PhysicsComponent? physics = null)
     {
         if (!Resolve(uid, ref comp) || !Resolve(uid, ref physics, false) || physics.FixturesMass <= 0)
@@ -14,5 +19,19 @@ public abstract partial class SharedTemperatureSystem : EntitySystem
         }
 
         return comp.SpecificHeat * physics.FixturesMass;
+    }
+}
+
+public sealed class OnTemperatureChangeEvent : EntityEventArgs
+{
+    public float CurrentTemperature { get; }
+    public float LastTemperature { get; }
+    public float TemperatureDelta { get; }
+
+    public OnTemperatureChangeEvent(float current, float last, float delta)
+    {
+        CurrentTemperature = current;
+        LastTemperature = last;
+        TemperatureDelta = delta;
     }
 }
